+++ packages/gasket-cli/src/scaffold/actions/index.js
@@ -1,13 +1,13 @@
 
 const mkDir = require('./mkdir');
 const loadPreset = require('./load-preset');
+const cliVersion = require('./cli-version');
 const globalPrompts = require('./global-prompts');
 const setupPkg = require('./setup-pkg');
 const writePkg = require('./write-pkg');
 const installModules = require('./install-modules');
 const linkModules = require('./link-modules');
 const writeGasketConfig = require('./write-gasket-config');
-const gitInit = require('./git-init');
 const loadPkgForDebug = require('./load-pkg-for-debug');
 const promptHooks = require('./prompt-hooks');
 const createHooks = require('./create-hooks');
@@ -19,13 +19,13 @@ const printReport = require('./print-report');
 module.exports = {
   mkDir,
   loadPreset,
+  cliVersion,
   globalPrompts,
   setupPkg,
   writePkg,
   installModules,
   linkModules,
   writeGasketConfig,
-  gitInit,
   loadPkgForDebug,
   promptHooks,
   createHooks,