+++ packages/gasket-cli/test/unit/commands/create.test.js
@@ -18,13 +18,13 @@ describe('create', function () {
     actionStubs = {
       mkDir: sandbox.stub(),
       loadPreset: sandbox.stub(),
+      cliVersion: sandbox.stub(),
       globalPrompts: sandbox.stub(),
       setupPkg: sandbox.stub(),
       writePkg: sandbox.stub(),
       installModules: sandbox.stub(),
       linkModules: sandbox.stub(),
       writeGasketConfig: sandbox.stub(),
-      gitInit: sandbox.stub(),
       loadPkgForDebug: sandbox.stub(),
       promptHooks: sandbox.stub(),
       createHooks: sandbox.stub(),
@@ -76,7 +76,6 @@ describe('create', function () {
     assume(actionStubs.writePkg).is.called();
     assume(actionStubs.installModules).is.called();
     assume(actionStubs.linkModules).is.called();
-    assume(actionStubs.gitInit).is.called();
     assume(actionStubs.postCreateHooks).is.called();
   });