+++ packages/gasket-manifest-plugin/lib/base-config.js
@@ -1,0 +1,10 @@
+/**
+ * Default values for manifest.json
+ * @type {Object}
+ */
+module.exports =  {
+  start_url: '/?source=pwa',
+  scope: '/',
+  display: 'standalone',
+  path: '/manifest.json'
+};