+++ packages/gasket-service-worker-plugin/__mocks__/uglify-js.js
@@ -1,0 +1,8 @@
+const minifyStub = jest.fn(() => ({
+  code: '() => {}'
+}));
+module.exports = jest.fn(() => ({
+  minify: minifyStub
+}));
+
+module.exports.minify = minifyStub;