+++ packages/gasket-service-worker-plugin/jest.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  collectCoverageFrom: [
+    'lib/**/*.js',
+    '!lib/**/*template.js'
+  ]
+};