+++ packages/gasket-utils/lib/index.js
@@ -1,7 +1,9 @@
 const tryRequire = require('./try-require');
 const applyEnvironmentOverrides = require('./apply-env-overrides');
+const runShellCommand = require('./run-shell-command');
 
 module.exports = {
   tryRequire,
-  applyEnvironmentOverrides
+  applyEnvironmentOverrides,
+  runShellCommand
 };