+++ packages/gasket-express-plugin/CHANGELOG.md
@@ -1,7 +1,13 @@
 # CHANGELOG
 
+### 2.0.0
+
+- [#23] Dismantling core-plugin
+
 ### 1.0.0
 
 - Renaming express-create to expressCreate
 - Separate express plugins from core
 - Initial release.
+
+[#23]: https://github.com/godaddy/gasket/pull/23