+++ packages/gasket-nextjs-plugin/CHANGELOG.md
@@ -1,5 +1,17 @@
 # CHANGELOG
 
+### 2.0.2
+
+- [#33] Add required dependencies during `gasket create`
+
+### 2.0.1
+
+- [#32] Do not build with `gasket local` command
+
+### 2.0.0
+
+- [#23] Dismantling core-plugin
+
 ### 1.1.1
 
 - Fix for missing dependency during create command
@@ -9,3 +21,7 @@
 - Renaming hooks and removing the `webpack` hook
 - Separate next plugins from core
 - Initial release.
+
+[#23]: https://github.com/godaddy/gasket/pull/23
+[#32]: https://github.com/godaddy/gasket/pull/32
+[#33]: https://github.com/godaddy/gasket/pull/33