+++ packages/gasket-utils/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+### 1.1.0
+
+- Adds functions:
+  - `runShellCommand`
+
 ### 1.0.0
 
 - Initial implementation. Adds: